Export FilesetID in fileset metadata
Some use cases, e.g. downloading files, require a fileset ID. Therefore
the METS metadata (and binary fileset metadata) should output
a filesetID if this is available.

At present, there is no need to crosswalk this value, as it should only be
read, not set on ingest. Therefore it can be automatically generated
when the relevant metadata is viewed, but not otherwise processed.

This update adds the output in the form ora:filesetId, as documented
in the sample METS documentation (and sample binary fileset metadata
example).